Add files via upload
This project used ocean water temperature data obtained from Ocean Observatories Initiative (OOI). The project objective was to compare the average temperature during each month of the year 2017 to determine how much the temperature changed during the year. The data was collected using a CTD at the same location (45.8305°N, 129.7535°W) at a depth of 190 meters below the surface. The data was analyzed over a 24 hour period during the first day of each month. Data for August 2017 was unavailable. 

The minimum average temperature was 44.579012 °F. The maximum average temperature was 45.6986894 °F. The total temperature change was approximately 1.1 °F during 2017. The data was obtained 190 meters below the surface, which is near the end of the Epipelagic (sunlight) Zone.